requirements.txt
asgiref==3.3.1
Django==3.1.8
django-filter==2.4.0
djangorestframework==3.12.2
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
pyyaml
django-rest-swagger
httpie